example of snapshot and api testing
