+++ joplin/api/snapshots/snap_test_api.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# snapshottest: v1 - https://goo.gl/zC4yUc
+from __future__ import unicode_literals
+
+from snapshottest import Snapshot
+
+
+snapshots = Snapshot()
+
+snapshots['test_service 1'] = {
+    'data': {
+        'allServicePages': {
+            'edges': [
+            ]
+        }
+    }
+}
